+++ src/components/order/OrderItemsList.jsx
@@ -1,8 +1,8 @@
-import CartOrderItem from "./CartOrderItem";
+import CartOrderItem from "../CartOrderItem";
 
 function OrderItemsList({ items }) {
   return (
-    <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-6">
+    <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-2 gap-4">
       {items.map((item) => (
         <CartOrderItem key={item.product_id} item={item} />
       ))}
